Move movement of functions into serialtbx
Moving out of cspad_cbf_tbx:
- Into serialtbx.detector: basis, center
- Into serialtbx.detector.cspad: read_slac_metrology
- Into serialtbx.detector.xtc: basis_from_geo
