+++ xfel/cftbx/detector/cspad_cbf_tbx.py
@@ -22,7 +22,8 @@ except AttributeError:
 asic_dimension = (194,185)
 asic_gap = 3
 pixel_size = 0.10992
-from serialtbx.detector.cspad import cspad_saturated_value, cspad_min_trusted_value
+from serialtbx.detector import basis, center
+from serialtbx.detector.cspad import cspad_saturated_value, cspad_min_trusted_value, read_slac_metrology
 
 def get_psana_corrected_data(psana_det, evt, use_default=False, dark=True, common_mode=None, apply_gain_mask=True,
                              gain_mask_value=None, per_pixel_gain=False, gain_mask=None, additional_gain_factor=None):
@@ -123,199 +124,6 @@ def angle_and_axis(basis):
   @param params metrology object"""
   q = matrix.col(basis.orientation).normalize()
   return q.unit_quaternion_as_axis_and_angle(deg=True)
-
-def center(coords):
-  """ Returns the average of a list of vectors
-  @param coords List of vectors to return the center of
-  """
-  for c in coords:
-    if 'avg' not in locals():
-      avg = c
-    else:
-      avg += c
-  return avg / len(coords)
-
-class basis(object):
-  """ Bucket for detector element information """
-  def __init__(self, orientation = None, translation = None, panelgroup = None, homogenous_transformation = None, name = None):
-    """
-    Provide only orientation + translation or a panelgroup or a homogenous_transformation.
-
-    @param orientation rotation in the form of a quarternion
-    @param translation vector translation in relation to the parent frame
-    @param panelgroup dxtbx panelgroup object whose local d matrix will represent the
-    basis shift
-    @param homogenous_transformation 4x4 matrix.sqr object representing a translation
-    and a rotation. Must not also contain a scale as this won't be decomposed properly.
-    @param name optional name for this basis shift
-    """
-    self.include_translation = True
-    self.name = name
-
-    if panelgroup is not None:
-      d_mat = panelgroup.get_local_d_matrix()
-      fast = matrix.col((d_mat[0],d_mat[3],d_mat[6])).normalize()
-      slow = matrix.col((d_mat[1],d_mat[4],d_mat[7])).normalize()
-      orig = matrix.col((d_mat[2],d_mat[5],d_mat[8]))
-
-      v3 = fast.cross(slow).normalize()
-
-      r3 = matrix.sqr((fast[0],slow[0],v3[0],
-                       fast[1],slow[1],v3[1],
-                       fast[2],slow[2],v3[2]))
-
-      self.orientation = r3.r3_rotation_matrix_as_unit_quaternion()
-      self.translation = orig
-
-      if not self.name:
-        self.name = panelgroup.get_name()
-
-    elif orientation is not None or translation is not None:
-      assert orientation is not None and translation is not None
-      self.orientation = orientation
-      self.translation = translation
-
-    else:
-      # Decompose the homegenous transformation assuming no scale factors were used
-      h = homogenous_transformation
-      self.orientation = matrix.sqr((h[0],h[1],h[2],
-                                     h[4],h[5],h[6],
-                                     h[8],h[9],h[10])).r3_rotation_matrix_as_unit_quaternion()
-      self.translation = matrix.col((h[3],
-                                     h[7],
-                                     h[11]))
-      assert h[12] == h[13] == h[14] == 0 and h[15] == 1
-
-  def as_homogenous_transformation(self):
-    """ Returns this basis change as a 4x4 transformation matrix in homogenous coordinates"""
-    r3 = self.orientation.normalize().unit_quaternion_as_r3_rotation_matrix()
-    return matrix.sqr((r3[0],r3[1],r3[2],self.translation[0],
-                       r3[3],r3[4],r3[5],self.translation[1],
-                       r3[6],r3[7],r3[8],self.translation[2],
-                       0,0,0,1))
-
-  def __mul__(self, other):
-    """ Use homogenous matrices to multiply bases together """
-    if hasattr(other, 'as_homogenous_transformation'):
-      return basis(homogenous_transformation = self.as_homogenous_transformation() * other.as_homogenous_transformation())
-    elif hasattr(other, 'n'):
-      if other.n == (3,1):
-        b = matrix.col((other[0], other[1], other[2], 1))
-      elif other.n == (4,1):
-        b = other
-      else:
-        raise TypeError(b, "Incompatible matrices")
-      p = self.as_homogenous_transformation() * b
-      if other.n == (3,1):
-        return matrix.col(p[0:3])
-      else:
-        return p
-    else:
-      raise TypeError(b)
-
-
-def basis_from_geo(geo, use_z = True):
-  """ Given a psana GeometryObject, construct a basis object """
-  rotx = matrix.col((1,0,0)).axis_and_angle_as_r3_rotation_matrix(
-    geo.rot_x + geo.tilt_x, deg=True)
-  roty = matrix.col((0,1,0)).axis_and_angle_as_r3_rotation_matrix(
-    geo.rot_y + geo.tilt_y, deg=True)
-  rotz = matrix.col((0,0,1)).axis_and_angle_as_r3_rotation_matrix(
-    geo.rot_z + geo.tilt_z, deg=True)
-
-  rot = (rotx*roty*rotz).r3_rotation_matrix_as_unit_quaternion()
-
-  if use_z:
-    trans = matrix.col((geo.x0/1000, geo.y0/1000, geo.z0/1000))
-  else:
-    trans = matrix.col((geo.x0/1000, geo.y0/1000, 0))
-
-  return basis(orientation = rot, translation = trans)
-
-def read_slac_metrology(path = None, geometry = None, plot=False, include_asic_offset=False):
-  if path is None and geometry is None:
-    raise Sorry("Need to provide a geometry object or a path to a geometry file")
-
-  if path is not None and geometry is not None:
-    raise Sorry("Cannot provide a geometry object and a geometry file. Ambiguous")
-
-  if geometry is None:
-    try:
-      from PSCalib.GeometryAccess import GeometryAccess
-      geometry = GeometryAccess(path)
-    except Exception as e:
-      raise Sorry("Can't parse this metrology file")
-
-  metro = {}
-  pixel_size = geometry.get_pixel_scale_size()/1000
-  null_ori = matrix.col((0,0,1)).axis_and_angle_as_unit_quaternion(0, deg=True)
-
-  # collapse any transformations above those of the quadrants into one X/Y offset,
-  # but don't keep Z transformations, as those come from the XTC stream
-  root = geometry.get_top_geo()
-  root_basis = basis_from_geo(root, use_z=False)
-  while len(root.get_list_of_children()) != 4 and len(root.get_list_of_children()) != 32:
-    assert len(root.get_list_of_children()) == 1
-    root = root.get_list_of_children()[0]
-    root_basis *= basis_from_geo(root, use_z=False)
-
-  metro[(0,)] = root_basis
-
-  def add_sensor(quad_id, sensor_id, sensor):
-    metro[(0,quad_id,sensor_id)] = basis_from_geo(sensor)
-
-    x, y, z = sensor.get_pixel_coords()
-    x/=1000; y/=1000; z/=1000
-    assert x.shape == y.shape == z.shape
-    sensor_px_slow = x.shape[0]
-    sensor_px_fast = x.shape[1]
-    assert sensor_px_fast % 2 == 0
-
-    a0ul = sul = matrix.col((x[0,0],y[0,0],z[0,0]))
-    a1ur = sur = matrix.col((x[0,sensor_px_fast-1],y[0,sensor_px_fast-1],z[0,sensor_px_fast-1]))
-    a1lr = slr = matrix.col((x[sensor_px_slow-1,sensor_px_fast-1],y[sensor_px_slow-1,sensor_px_fast-1],z[sensor_px_slow-1,sensor_px_fast-1]))
-    a0ll = sll = matrix.col((x[sensor_px_slow-1,0],y[sensor_px_slow-1,0],z[sensor_px_slow-1,0]))
-
-    a0ur = matrix.col((x[0,sensor_px_fast//2-1],y[0,sensor_px_fast//2-1],z[0,sensor_px_fast//2-1]))
-    a0lr = matrix.col((x[sensor_px_slow-1,sensor_px_fast//2-1],y[sensor_px_slow-1,sensor_px_fast//2-1],z[sensor_px_slow-1,sensor_px_fast//2-1]))
-
-    a1ul = matrix.col((x[0,sensor_px_fast//2],y[0,sensor_px_fast//2],z[0,sensor_px_fast//2]))
-    a1ll = matrix.col((x[sensor_px_slow-1,sensor_px_fast//2],y[sensor_px_slow-1,sensor_px_fast//2],z[sensor_px_slow-1,sensor_px_fast//2]))
-
-    sensor_center = center([sul,sur,slr,sll])
-    asic0_center = center([a0ul,a0ur,a0lr,a0ll])
-    asic1_center = center([a1ul,a1ur,a1lr,a1ll])
-
-    asic_trans0 = (asic0_center-sensor_center).length()
-    asic_trans1 = (asic1_center-sensor_center).length()
-
-    if include_asic_offset:
-      rotated_ori = matrix.col((1,0,0)).axis_and_angle_as_unit_quaternion(180.0, deg=True)
-      offset_fast = -pixel_size*((sensor_px_fast) / 4) # 4 because sensor_px_fast is for sensor
-      offset_slow = +pixel_size*((sensor_px_slow) / 2) # Sensor is divided into 2 only in fast direction
-      metro[(0,quad_id,sensor_id,0)] = basis(orientation=rotated_ori,translation=matrix.col((-asic_trans0,0,0)))
-      metro[(0,quad_id,sensor_id,1)] = basis(orientation=rotated_ori,translation=matrix.col((+asic_trans1,0,0)))
-      metro[(0,quad_id,sensor_id,0)].translation += matrix.col((offset_fast, offset_slow, 0))
-      metro[(0,quad_id,sensor_id,1)].translation += matrix.col((offset_fast, offset_slow, 0))
-    else:
-      metro[(0,quad_id,sensor_id,0)] = basis(orientation=null_ori,translation=matrix.col((-asic_trans0,0,0)))
-      metro[(0,quad_id,sensor_id,1)] = basis(orientation=null_ori,translation=matrix.col((+asic_trans1,0,0)))
-
-  if len(root.get_list_of_children()) == 4:
-    for quad_id, quad in enumerate(root.get_list_of_children()):
-      metro[(0,quad_id)] = basis_from_geo(quad)
-      for sensor_id, sensor in enumerate(quad.get_list_of_children()):
-        add_sensor(quad_id, sensor_id, sensor)
-  elif len(root.get_list_of_children()) == 32:
-    for quad_id in range(4):
-      metro[(0,quad_id)] = basis(orientation = null_ori, translation = matrix.col((0,0,0)))
-      sensors = root.get_list_of_children()
-      for sensor_id in range(8):
-        add_sensor(quad_id, sensor_id, sensors[quad_id*4+sensor_id])
-  else:
-    assert False
-
-  return metro
 
 def get_calib_file_path(env, address, run):
   """ Findes the path to the SLAC metrology file stored in a psana environment
